ChiDT/ refs logic solve money and return to api

+++ Backend/src/main/java/vn/edu/ptit/sqa/model/reportInfor/DebtReportResponse.java
@@ -9,6 +9,6 @@ import java.util.List;
 public class DebtReportResponse {
     private DebtReportDTO debtReportDTO;
     private List<DebtCustomerDTO> debtCustomerList;
-    private Long totalDebtNumber;
+    private float totalDebtNumber;
     private PageDto pageDto;
 }
